+++ etils/epy/__init__.py
@@ -21,6 +21,7 @@ from etils.epy.contextlib import ContextManager
 from etils.epy.env_utils import is_notebook
 from etils.epy.itertools import groupby
 from etils.epy.itertools import zip_dict
+from etils.epy.py_utils import frozen
 from etils.epy.py_utils import issubclass_ as issubclass  # pylint: disable=redefined-builtin
 from etils.epy.py_utils import StrEnum
 from etils.epy.reraise_utils import maybe_reraise